requirements.txt
asgiref
django
Pillow
sqlparse
tzdata
gunicorn
whitenoise
pytesseract
